Move process readers to new file
