+++ src/model.rs
@@ -3,7 +3,8 @@
 use crate::agent::{read_agents, Agent};
 use crate::commodity::{read_commodities, Commodity};
 use crate::input::*;
-use crate::process::{read_processes, Process};
+use crate::process::Process;
+use crate::process_readers::read_processes;
 use crate::region::{read_regions, Region};
 use crate::time_slice::{read_time_slice_info, TimeSliceInfo};
 use anyhow::{ensure, Context, Result};
